# contracts/openapi.json
{
  "components": {
    "schemas": {
      "AnswerStyle": {
        "enum": [
          "concise",
          "detailed"
        ],
        "title": "AnswerStyle",
        "type": "string"
      },
...
      "Effort": {
        "enum": [
          "low",
          "medium",
          "high"
        ],
        "title": "Effort",
        "type": "string"
      },
      "ErrorBody": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "details": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Details",
            "type": "array"
          },
          "message": {
            "description": "Developer-facing text for logs. The UI translates `code`.",
            "title": "Message",
            "type": "string"
          },
          "params": {
            "additionalProperties": true,
            "title": "Params",
            "type": "object"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "retry_after": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Retry After"
          },
          "retryable": {
            "title": "Retryable",
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message",
          "retryable",
          "request_id"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorCode": {
        "enum": [
          "VALIDATION_ERROR",
          "NOT_FOUND",
          "METHOD_NOT_ALLOWED",
          "UNAUTHORIZED_CLIENT",
          "SERVICE_STARTING",
          "INTERNAL_ERROR",
          "REQUEST_TOO_LARGE",
          "RATE_LIMITED",
          "UPLOAD_TOO_LARGE",
          "UNSUPPORTED_TYPE",
          "FILE_CONTENT_MISMATCH",
          "EMPTY_FILE",
          "DUPLICATE_DOCUMENT",
          "STORAGE_QUOTA",
          "STORAGE_FULL",
          "UPLOAD_INCOMPLETE",
          "URL_INVALID",
          "URL_BLOCKED",
          "URL_UNREACHABLE",
          "URL_TIMEOUT",
          "URL_TOO_LARGE",
          "URL_UNSUPPORTED_TYPE",
          "DOCUMENT_NOT_READY",
          "DOCUMENT_FILE_MISSING",
          "DELETE_FAILED",
          "RANGE_NOT_SATISFIABLE",
          "PDF_ENCRYPTED",
          "PDF_CORRUPT",
          "PDF_NO_TEXT",
          "PDF_TOO_MANY_PAGES",
          "TEXT_ENCODING_UNSUPPORTED",
          "DOCUMENT_EMPTY",
          "DOCUMENT_TOO_LONG",
          "PROCESSING_TIMEOUT",
          "PROCESSING_FAILED",
          "PROCESSING_INTERRUPTED",
          "MALWARE_DETECTED",
          "MALWARE_SCAN_FAILED",
          "CHAT_NOT_FOUND",
          "CHAT_BUSY",
          "CHAT_LIMIT",
          "MESSAGE_LIMIT",
          "MESSAGE_NOT_LATEST",
          "DUPLICATE_REQUEST",
          "CONCURRENCY_LIMIT",
          "NO_DOCUMENTS",
          "DOCUMENTS_NOT_READY",
          "QUESTION_EMPTY",
          "QUESTION_TOO_LONG",
          "MODEL_NOT_ALLOWED",
          "COMPARE_SAME_MODEL",
          "TOKEN_BUDGET_EXCEEDED",
          "LLM_AUTH",
          "API_KEY_INVALID",
          "LLM_KEY_NEEDS_WORKSPACE",
          "LLM_BILLING",
          "LLM_FORBIDDEN",
          "MODEL_UNAVAILABLE",
          "LLM_RATE_LIMITED",
          "LLM_OVERLOADED",
          "LLM_UNAVAILABLE",
          "LLM_TIMEOUT",
          "LLM_UNREACHABLE",
          "LLM_BAD_REQUEST",
          "LLM_CONTEXT_TOO_LARGE",
          "LLM_EMPTY_ANSWER"
        ],
        "title": "ErrorCode",
        "type": "string",
        "x-error-specs": {
          "API_KEY_INVALID": {
            "retryable": false,
            "status": 422
          },
          "CHAT_BUSY": {
            "retryable": true,
            "status": 409
          },
          "CHAT_LIMIT": {
            "retryable": false,
            "status": 409
          },
          "CHAT_NOT_FOUND": {
            "retryable": false,
            "status": 404
          },
          "COMPARE_SAME_MODEL": {
            "retryable": false,
            "status": 422
          },
          "CONCURRENCY_LIMIT": {
            "retryable": true,
            "status": 429
          },
          "DELETE_FAILED": {
            "retryable": true,
            "status": 500
          },
          "DOCUMENTS_NOT_READY": {
            "retryable": true,
            "status": 409
          },
          "DOCUMENT_EMPTY": {
            "retryable": false,
            "status": 422
          },
          "DOCUMENT_FILE_MISSING": {
            "retryable": false,
            "status": 410
          },
          "DOCUMENT_NOT_READY": {
            "retryable": true,
            "status": 409
          },
          "DOCUMENT_TOO_LONG": {
            "retryable": false,
            "status": 422
          },
          "DUPLICATE_DOCUMENT": {
            "retryable": false,
            "status": 409
          },
          "DUPLICATE_REQUEST": {
            "retryable": false,
            "status": 409
          },
          "EMPTY_FILE": {
            "retryable": false,
            "status": 422
          },
          "FILE_CONTENT_MISMATCH": {
            "retryable": false,
            "status": 415
          },
          "INTERNAL_ERROR": {
            "retryable": true,
            "status": 500
          },
          "LLM_AUTH": {
            "retryable": false,
            "status": 503
          },
          "LLM_BAD_REQUEST": {
            "retryable": false,
            "status": 500
          },
          "LLM_BILLING": {
            "retryable": false,
            "status": 503
          },
          "LLM_CONTEXT_TOO_LARGE": {
            "retryable": false,
            "status": 500
          },
          "LLM_EMPTY_ANSWER": {
            "retryable": true,
            "status": 502
          },
          "LLM_FORBIDDEN": {
            "retryable": false,
            "status": 503
          },
          "LLM_KEY_NEEDS_WORKSPACE": {
            "retryable": false,
            "status": 422
          },
          "LLM_OVERLOADED": {
            "retryable": true,
            "status": 503
          },
          "LLM_RATE_LIMITED": {
            "retryable": true,
            "status": 503
          },
          "LLM_TIMEOUT": {
            "retryable": true,
            "status": 504
          },
          "LLM_UNAVAILABLE": {
            "retryable": true,
            "status": 502
          },
          "LLM_UNREACHABLE": {
            "retryable": true,
            "status": 502
          },
          "MALWARE_DETECTED": {
            "retryable": false,
            "status": 422
          },
          "MALWARE_SCAN_FAILED": {
            "retryable": true,
            "status": 422
          },
          "MESSAGE_LIMIT": {
            "retryable": false,
            "status": 409
          },
          "MESSAGE_NOT_LATEST": {
            "retryable": false,
            "status": 409
          },
          "METHOD_NOT_ALLOWED": {
            "retryable": false,
            "status": 405
          },
          "MODEL_NOT_ALLOWED": {
            "retryable": false,
            "status": 422
          },
          "MODEL_UNAVAILABLE": {
            "retryable": false,
            "status": 503
          },
          "NOT_FOUND": {
            "retryable": false,
            "status": 404
          },
          "NO_DOCUMENTS": {
            "retryable": false,
            "status": 409
          },
          "PDF_CORRUPT": {
            "retryable": false,
            "status": 422
          },
          "PDF_ENCRYPTED": {
            "retryable": false,
            "status": 422
          },
          "PDF_NO_TEXT": {
            "retryable": false,
            "status": 422
          },
          "PDF_TOO_MANY_PAGES": {
            "retryable": false,
            "status": 422
          },
          "PROCESSING_FAILED": {
            "retryable": true,
            "status": 500
          },
          "PROCESSING_INTERRUPTED": {
            "retryable": true,
            "status": 500
          },
          "PROCESSING_TIMEOUT": {
            "retryable": true,
            "status": 500
          },
          "QUESTION_EMPTY": {
            "retryable": false,
            "status": 422
          },
          "QUESTION_TOO_LONG": {
            "retryable": false,
            "status": 422
          },
          "RANGE_NOT_SATISFIABLE": {
            "retryable": false,
            "status": 416
          },
          "RATE_LIMITED": {
            "retryable": true,
            "status": 429
          },
          "REQUEST_TOO_LARGE": {
            "retryable": false,
            "status": 413
          },
          "SERVICE_STARTING": {
            "retryable": true,
            "status": 503
          },
          "STORAGE_FULL": {
            "retryable": false,
            "status": 507
          },
          "STORAGE_QUOTA": {
            "retryable": false,
            "status": 409
          },
          "TEXT_ENCODING_UNSUPPORTED": {
            "retryable": false,
            "status": 422
          },
          "TOKEN_BUDGET_EXCEEDED": {
            "retryable": true,
            "status": 429
          },
          "UNAUTHORIZED_CLIENT": {
            "retryable": false,
            "status": 401
          },
          "UNSUPPORTED_TYPE": {
            "retryable": false,
            "status": 415
          },
          "UPLOAD_INCOMPLETE": {
            "retryable": true,
            "status": 400
          },
          "UPLOAD_TOO_LARGE": {
            "retryable": false,
            "status": 413
          },
          "URL_BLOCKED": {
            "retryable": false,
            "status": 422
          },
          "URL_INVALID": {
            "retryable": false,
            "status": 422
          },
          "URL_TIMEOUT": {
            "retryable": true,
            "status": 504
          },
          "URL_TOO_LARGE": {
            "retryable": false,
            "status": 413
          },
          "URL_UNREACHABLE": {
            "retryable": true,
            "status": 502
          },
          "URL_UNSUPPORTED_TYPE": {
            "retryable": false,
            "status": 415
          },
          "VALIDATION_ERROR": {
            "retryable": false,
            "status": 422
          }
        }
      },
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "Locale": {
        "enum": [
          "de",
          "en"
        ],
        "title": "Locale",
        "type": "string"
      },
...
      "PreferencesBody": {
        "additionalProperties": false,
        "description": "Always the whole object.",
        "properties": {
          "compare_models": {
            "items": {
              "type": "string"
            },
            "maxItems": 2,
            "minItems": 2,
            "title": "Compare Models",
            "type": "array"
          },
          "default_model": {
            "maxLength": 100,
            "title": "Default Model",
            "type": "string"
          },
          "effort": {
            "$ref": "#/components/schemas/Effort",
            "description": "Answer mode: fast (low), balanced, thorough (high)."
          },
          "locale": {
            "$ref": "#/components/schemas/Locale"
          },
          "name": {
            "description": "Only for the greeting, never sent to the model. Cleaned on save.",
            "maxLength": 40,
            "title": "Name",
            "type": "string"
          },
          "onboarded": {
            "title": "Onboarded",
            "type": "boolean"
          },
          "retention_days": {
            "anyOf": [
              {
                "maximum": 3650.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Delete chats and documents after this many days; 0 is off. Until chosen, the installation default (RETENTION_DAYS). Omitted on PUT: stays as it is (clients written before this field existed).",
            "title": "Retention Days"
          },
          "style": {
            "$ref": "#/components/schemas/AnswerStyle"
          },
          "theme": {
            "$ref": "#/components/schemas/Theme"
          }
        },
        "required": [
          "locale",
          "theme",
          "name",
          "default_model",
          "effort",
          "style",
          "compare_models",
          "onboarded"
        ],
        "title": "PreferencesBody",
        "type": "object"
      },
...
      "Theme": {
        "enum": [
          "light",
          "dark",
          "system"
        ],
        "title": "Theme",
        "type": "string"
      },
...
      "UsageMonthOut": {
        "properties": {
          "cost_usd": {
            "title": "Cost Usd",
            "type": "number"
          },
          "requests": {
            "title": "Requests",
            "type": "integer"
          }
        },
        "required": [
          "cost_usd",
          "requests"
        ],
        "title": "UsageMonthOut",
        "type": "object"
      },
...
      "UsageTodayOut": {
        "properties": {
          "budget_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "DAILY_BUDGET_USD; null is off.",
            "title": "Budget Usd"
          },
          "cost_usd": {
            "title": "Cost Usd",
            "type": "number"
          },
          "input_tokens": {
            "title": "Input Tokens",
            "type": "integer"
          },
          "output_tokens": {
            "title": "Output Tokens",
            "type": "integer"
          },
          "requests": {
            "title": "Requests",
            "type": "integer"
          }
        },
        "required": [
          "cost_usd",
          "requests",
          "input_tokens",
          "output_tokens",
          "budget_usd"
        ],
        "title": "UsageTodayOut",
        "type": "object"
      },
      "WorkspaceOut": {
        "properties": {
          "retention_days": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "From the preferences; null: nothing is deleted automatically.",
            "title": "Retention Days"
          },
          "stats": {
            "$ref": "#/components/schemas/WorkspaceStatsOut"
          },
          "usage_month": {
            "$ref": "#/components/schemas/UsageMonthOut",
            "description": "Since the first of the current UTC month: what this app spent at Anthropic."
          },
          "usage_today": {
            "$ref": "#/components/schemas/UsageTodayOut",
            "description": "The current UTC day."
          }
        },
        "required": [
          "stats",
          "usage_today",
          "usage_month",
          "retention_days"
        ],
        "title": "WorkspaceOut",
        "type": "object"
      },
      "WorkspaceStatsOut": {
        "properties": {
          "chats": {
            "title": "Chats",
            "type": "integer"
          },
          "documents": {
            "title": "Documents",
            "type": "integer"
          },
          "documents_bytes": {
            "description": "Size of the original files.",
            "title": "Documents Bytes",
            "type": "integer"
          },
          "storage_bytes": {
            "description": "Everything on disk: files, search index, database.",
            "title": "Storage Bytes",
            "type": "integer"
          }
        },
        "required": [
          "documents",
          "chats",
          "documents_bytes",
          "storage_bytes"
        ],
        "title": "WorkspaceStatsOut",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Siteco Document Chat API",
    "version": "contract"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/preferences": {
      "get": {
        "description": "The defaults (`onboarded` false) until the setup ran.",
        "operationId": "get_preferences_api_preferences_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreferencesBody"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Preferences",
        "tags": [
          "workspace"
        ]
      },
      "put": {
        "description": "The whole object. 422 `MODEL_NOT_ALLOWED` for a model that is not offered.",
        "operationId": "put_preferences_api_preferences_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreferencesBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreferencesBody"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Put Preferences",
        "tags": [
          "workspace"
        ]
      }
    },
...
    "/api/workspace": {
      "delete": {
        "description": "Deletes every document, chat and message. Running answers are stopped first.",
        "operationId": "delete_workspace_api_workspace_delete",
        "parameters": [
          {
            "description": "Also forget name and settings; the setup shows again.",
            "in": "query",
            "name": "reset_preferences",
            "required": false,
            "schema": {
              "default": false,
              "description": "Also forget name and settings; the setup shows again.",
              "title": "Reset Preferences",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Delete Workspace",
        "tags": [
          "workspace"
        ]
      },
      "get": {
        "operationId": "get_workspace_api_workspace_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceOut"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Workspace",
        "tags": [
          "workspace"
        ]
      }
    },
...
  }
}